+++ src/eutxo/mod.rs
@@ -1,7 +1,5 @@
 pub mod btc;
-pub mod eutxo_model;
 pub mod eutxo_block_monitor;
-pub mod eutxo_codec_block;
-pub mod eutxo_codec_tx;
 pub mod eutxo_codec_utxo;
-pub mod eutxo_service;
+pub mod eutxo_model;
+pub mod eutxo_tx_service;
