+++ src/main.rs
@@ -1,9 +1,9 @@
 use ci::api::ChainLinker;
+use ci::block_service::BlockService;
 use ci::eutxo::btc::btc_chain_linker::BtcChainLinker;
-use ci::eutxo::btc::{btc_client::BtcClient, btc_processor::BtcProcessor};
 use ci::eutxo::eutxo_block_monitor::EuBlockMonitor;
 use ci::eutxo::eutxo_model::{self, EuTx};
-use ci::eutxo::eutxo_service::EuService;
+use ci::eutxo::eutxo_tx_service::EuTxService;
 use ci::indexer::Indexer;
 use ci::settings::AppConfig;
 use ci::storage::Storage;
@@ -32,17 +32,21 @@ async fn main() -> Result<(), std::io::Error> {
                         eutxo_model::get_eutxo_column_families(),
                     ));
                     // let db_holder = Arc::new(DbHolder { db: Arc::new(db) });
-                    let service: Arc<EuService> = Arc::new(EuService::new());
+                    let tx_service: Arc<EuTxService> = Arc::new(EuTxService {});
+                    let block_service: Arc<BlockService<EuTx>> =
+                        Arc::new(BlockService::new(tx_service));
+
                     let chain_linker: Arc<
                         dyn ChainLinker<InTx = bitcoin::Transaction, OutTx = EuTx> + Send + Sync,
-                    > = Arc::new(BtcChainLinker {
-                        client: BtcClient::new(&api_host, &api_username, &api_password),
-                        processor: BtcProcessor {},
-                    });
+                    > = Arc::new(BtcChainLinker::new(&api_host, &api_username, &api_password));
                     let syncer = ChainSyncer::new(
                         Arc::clone(&chain_linker),
                         Arc::new(EuBlockMonitor::new(1000)),
-                        Arc::new(Indexer::new(db_holder, service, Arc::clone(&chain_linker))),
+                        Arc::new(Indexer::new(
+                            db_holder,
+                            block_service,
+                            Arc::clone(&chain_linker),
+                        )),
                     );
                     syncer.sync(tx_batch_size).await;
                     Ok(())
